requirements/base.txt
-c constraints.txt
pretty-midi==0.2.9
PyYAML==5.3.1
sinethesizer==0.2.9
